refactor Address, DomesticAddress, migrations
