+++ composer.json
@@ -18,7 +18,7 @@
         "php": "^7.4|^8.0",
         "dniccum/phone-number": "^1.1",
         "inspheric/nova-email-field": "^1.42",
-        "tipoff/authorization": "^2.3.2",
+        "tipoff/authorization": "^2.5.0",
         "tipoff/support": "^1.5.10"
     },
     "require-dev": {
