+++ tests/Unit/Models/AddressModelTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Tipoff\Addresses\Tests\Unit\Models;
 
 use Illuminate\Foundation\Testing\DatabaseTransactions;
-use Tipoff\Addresses\Models\Address;
+use Tipoff\Addresses\Models\DomesticAddress;
 use Tipoff\Addresses\Tests\TestCase;
 
 class AddressModelTest extends TestCase
@@ -15,7 +15,7 @@ class AddressModelTest extends TestCase
     /** @test */
     public function create()
     {
-        $model = Address::factory()->create();
+        $model = DomesticAddress::factory()->create();
         $this->assertNotNull($model);
     }
 }
